Renamed file
from test_huggingface_hub_bucketfs_model_transfer_sp.py
to test_model_utils.py

and refactored test to use UDF from test resources.
